requirements.txt
requests==2.31.0
pyyaml==6.0.1
pandas==2.1.4
python-dotenv==1.0.0
